Rewrote the structure for how BrowseActivity uses the classes that are inhariting from it.
There is an UpdateList method that the classes should call at the end of
theire onCreate method, this method will call the asyncUpdate method in
an async thread, in there the classes should populate the item list with
what it wants to have end up in the list and BrowseActivity will take
care of that part and generate the menus.

Modified all classes to follow this behavior

Added an Add method to FSActivity so that long-press and the menu there
will work for both files and folders.