StreamingService: Explicit try/catch & remove/reverse conditionals.
The point of this commit is to remove the deep layers of nested conditionals
inside showNotification(). This will make this member easier to read and
improve. I do this by removing unnecessary conditionals and reversing/returning
conditionals as required.

This commit also removes the try/catch block which encompassed the entire
method. If we crash MPDroid because we're not handling exceptions, we need to
handle them and not paper over the problem, IMHO.