StreamingService: Don't assume that album, artist or title exist.
MPDroid's old behaviour when adding the album, artist or title to the
notification panel is if it's null the notification panel would print the
string 'null' to the notification panel.

The new behaviour checks for the field first, then is either omitted, as is the
case with album or artist, or is substituted with filename in the case of
the title tag.