Add a setting to choose what artist tag is used for Artist list population. Closes #482
Handling libraries is hard. It's even harder when everybody has their own needs and ways of managing their library.

The removal of this setting has been controversial, but the old situation wasn't perfect either.
The setting now allow you to use the "artist" tag, "albumartist" or both.
There can be potential implications on the album list that follows an artist selection (since I don't propagate this option to AlbumsFragment), but since I'm not really sure how to approach this I'll address it (if needed : "both" is a great option for albums in my opinion) later.
Cache users are not affected by album handling since it works great anyway.