MPDStatusMonitor: Fix MPDStatusMonitor.trackChanged() for consume mode.
Turns out that MPDStatusMonitor.trackChanged() only worked when the number of
the song in relation to the current playlist queue changed (the songPos).
Without manual user intervention, the songPos will never iterate once consume
mode is entered. This patch changes the behaviour to use songId to trigger a
trackChanged() event instead. trackChanged() will continue to pass the
songPos to listeners.