StreamingService: At playlist completion, let onCompletion() handle stream end.
If the last song ends and the playlist is going to have a length of 0,
using onStateChange() will cut the final song off early due to buffering.
Instead, now check to see if the playlist is blank on stop and if so ignore
it and it will be handled in onComletion().