StreamingService: If playing and audio focus is lost, pause the player.
This is really no different than pulling the plug out of the earphone socket,
we pause on that, so we should pause on this as well. The only thing that is
substandard about this is this server might have other listeners. This is the
only sane thing to do in this case, though.