NotificationService: Make service persistent to option.
If NotificationMode is started, don't unselect it. Keep the service open,
unless the user opts to hit the close button.

Rather than update on track change, update on playlist change. This will be
more granular and it will include stream information updates unlike track
changes.

Finally, this patch turned out to be a small refactoring, so there are
many changes not targeted by the description.

For more information about how persistent notification works:
https://github.com/avuton/dmix/wiki/How-the-services-work-in-MPDroid