StreamingService: Prevent a MediaPlayer crash if stream halts during buffer.
MediaPlayer can cause an ANR or crash (depending on its state) if the stream
is stopped during buffer and the object is reset or released. Workaround that
by not resetting or releasing during buffering. The disadvantage to this is
that next start we may have some stale bits in the buffer which we'll hear
first, then it'll rebuffer.