NowPlayingFragment: Seed MPDStatus on connection or resume.
Prior to this, the status was kept up to date due to how many times
updateTrackInfoAsync() would get hit in the beginning. After this behaviour
was cleaned up the status would regularly be out of sync, this should
fix that.

The only two places the MPDStatus should need to be updated, outside
of the idle status monitor is connectionStateChanged() or onResume().