MPD.java: Improve Add, Remove & Play handling (from 6d0141a).
Before this patch, if the user selects Add, Remove & Play (ARP):

If consume mode is enabled, it chews up the first song (whoops!).

If random mode is enabled we would play the second song on the playlist queue,
which is unexpected behaviour, so we now send a next command in this case.

If Add & Play (not removing), and random mode is enabled, we would play:
(old playlist queue+1) which is flawed thinking, as in random mode it will
revert to the old playlist queue at any point afterwards, so, in this case,
play doesn't make sense, so we now toast the user that we don't Add & Play
in random mode.