MPDApplication: Dynamically check service status.
Setting & unsetting server status is not fullproof. Servers can crash, and
generally, go out of sync easily. This is approach is fullproof and clean. I
fully expect to replace this functionality as soon as I can figure out a more
robust method of accomplishing this.

This patch also removes the ApplicationState inner class as it is now unused.