StreamHandler: Don't extend Service, don't implement ServiceConnection.
Rationale:
* This is necessary, as Android expects the MediaPlayer to be attached to a
notification in order to take advantage of startForeground() to prevent
unexpected Service shutdown.

* This is also to allows access to the same AudioManager object; Android's
AudioManager crashed often when the services would take audio focus from each
other.

Implementation:
There are several things that had to happen at the same time to make this
patch work successfully which is why this patch is the size it is. Several
rewrites were attempted to make this happen gradually while keeping things
stable and working, after the sixth attempt this patch was created.

* This patch makes the service spawn it's own process. This significantly
reduces memory requirements of the service itself, allows more forgiving
instability, takes the service completely off the UI thread, and allows
MPDroid process killing without affecting the service (or visa versa).

* This patch adds helper functions to MPDApplication for start/stop and
server querying.

* This patch adds a Message handler class inside the MPDroid service to receive
and act on messages. Also adds debugging functions to help decode 'what'
messages.

* The MPDroidService now has a minimal startup. On startup very little will
load and is controlled by the new ServiceBinder class, currently implemented
in MPDApplication and started by the main activity to prevent a ServiceBinder
instance in the Service process. Service handlers will start dynamically as
necessary. One rationale for this change is the service must start at MPDroid
startup, minimally, to query server status. This is due to the Android
framework not having a way (prior to API 21) to know if a Service is even
running (even one owned by the main process), so deep Binder integration is
necessary for a concrete server implementation.

* The MPDroidService will also shut itself down when it's not being used.

* The MPDroid service does not use the Application instance now, instead it
uses the service instance, and only when it's absolutely necessary.

* This patch slightly changes Notification behaviour, it should not start on
Notification activation, it should only be active during play or the wind
down period.

* This patch adds stream stop/start to the ActionFire activity.

* In this patch, StreamHandler no longer extends Service, removes all binding
code, and relies on the main service for functionality.