AlbumCoverHandler: Don't recycle, just nullify and let GC do it's job.
I'm not sure exactly why but there are some times that Android can get out of
state and will try to parcel the albumCover after recycling. With Android 3.x+
it's really not worth the risk of crashing the service over it, so null it
and let GC do it's job.