service: Revamp the album cover listeners.
Notification always processed (scaling and decoded) it's own image using
RemoteViews.setImageViewUri() because RemoteViews.setImageViewBitmap() would
always crash due to recycling. This resulted in a noticable on UI thread lag.

This was caused by an errant shallow copy being sent to RemoteControlClient and
Notification, when one handler was done with the copy it would recycle for both
of them.