MusicList: Synchronize when iterating or using as a map pair.
According to {@code Collections.synchronizedList()}, "It is imperative that the
user manually synchronize on the returned list when iterating over it", so we
need to synchronize in that case.

We should also synchronize when using the map and the list as a pair, to
prevent us from getting unlucky in the case that the map gets modified by a
different thread at the same time.