dmix: MPDStatus (as returned from MPD) can no longer be null.
MPDStatus, as returned from MPD().getStatus() used to double as a status
update method. This caused all callers to require null checking and error
handling, in the case that a status refresh was required. This commit
uses StatusMontior callbacks to keep completely updated, from connection
to updates.

This also adds the method MPDStatus().isValid() to check status object
validity, since this was, typically, done as a null check when previously
required.