dmix: Implement unified equals/hashcode where overridden.
This patch adds a more readable Object.equals(object) implementation and
Object.hashCode(). This commit also adds hashCode() where it is missing.
This commit ensures the internal objects for computing hashCode() are final.
This commit removes equals()/hashCode() from CoverInfo, it's duplicative
of AlbumInfo. Finally, this moves StringsUtils to Tools.