MPDServerException: Add more robust and flexible error handling.
This commit extracts the error code, the command queue position, the command
sent and the ACK error message from the standard ACK error message for use
directly from the exception. This should allow for improved error handling and
cleaner error internationalization.

This commit also allows adding ACK codes to the command constructor to be
considered non-fatal, thus not throw exception.

This commit also adds the permission ACK to the password ACK which will not
allow retries of the current command.