Directory: Refactor refreshData().
This refactors refreshData() and getDir() as refresh(). This combines the two
as one method. This also fixes a few possible races. Finally, this fixes
issue #617. This bug was caused by quite an old regression where entry maps
were not cleared prior to updating, which is required because there is
no differential directory updating in the MPD protocol and we can't clear the
entries that no longer exist.