Album: Improve the SORT_BY_YEAR comparator.
Previously, if a 4 digit year was compared to a 8 digit date, the
comparison would always be less. This helper method will append 0s to the end
of the date if the length of the date is less than 8 and greater than 0. If
the date is greater than 8, the date will be trimmed to 8 digits.