MPDConnectionException: Replace with IOException.
MPDConnectionException should not extend MPD*Exception. Looking further into it
I came to the conclusion that it did nothing beyond what an IOException does.

This revision also moves MPDServerException to MPDException, as it was
redundant at this point as well.

It may come back in the future, if there is some interesting way to extend it,
but if it does, it will extend IOException.