MPDConnection: On no result / no exception, warn the user.
We received an automated crash report where a command result was not received,
and an exception was not emitted; MPDroid would crash (this is an unexpected
event). Unsure of how this could occur right now, but we have much more work
to take care of in the connection code and this will be fixed sooner or later,
if it is an actual bug. For now, warn the user, and don't crash.