MPDConnection: Throw IOException if no response and no error.
We've received a couple of automated crash reports which no exception was thrown
and no result was received. This is probably a bug of some sort. Having the user
crash isn't going to help, throw an IOException for now, until the bug is either
found or the code is replaced.