Album: Use a builder pattern for album construction, make immutable.
Album, uses many different types of construction and, it's getting harder to
construct, so a builder is natural for construction.

This patch:

* Removes setHasAlbumArtist(), this was required to make all fields immutable,
and work with the builder pattern.

* Moves the getAlbumDetails() method from the MPD classes to addAlbumDetails(),
this will be less confusing as it's a superset of addAlbumPaths().

* This removes the findYear parameter from addAlbumDetails(). The information
was already being retrieved, just not placed in the album object.