Music: Use smaller, more exact, constructors, fix NPE in getFullPath().
We received a crash report detailing a crash that occured due to a null
mFullPath field. This really shouldn't happen, but since it apparently does,
cover it with an if block like the rest of this class. This /could/ be fixed
with the constructor fix (detailed below); cover it up until possibly post-
release testing.

While looking into this bug, I also noticed the children classes for the Music
class are inexact constructors. For instance, we would construct the children
with getTitle() which is /not/ a getter. Use a better suited constructor and
change permissions so it won't happen again.