MPD: Don't send a command queue if nothing was added in add().
We received a automated crash response due to an attempt to send an empty
command queue. This was due to a user doing an add() with very specific
parameters which would negate every conditional on the way through and
never added to the command queue. For this case, I've added a conditional
which should be negated if there is an empty command queue at the end.