MPDStatusMonitor: Workaround an Android 5 bug.
First, let me start by saying this code change makes no logical sense, that I
can tell.

This issue is, intermittently, if state is changed quickly enough, the state
change wouldn't show.

After a few hours of bisection I traced this down to:
e0d9799:JMPDComm/src/main/java/org/a0z/mpd/MPDStatusMonitor.java:107

When this line is removed, and the code is replace with code that result in
the same code result (hard coded or otherwise) the state changes just as it
should. This is not a cloning issue, the state is not modified in MPDCommand
(and I could reproduce it post-cloning). Since I've only been able to reproduce
this in Android 5, so my wild assumption is that this is a bug in ART; but it's
easy enough to workaround that this patch is a trivial non-issue.