StreamHandler: If the stream is stopped during buffering, don't play.
This should fix an inconsistency bug where if a stream was playing, and the
stream would be stopped by the checkbox during buffering, the stream would
continue playing and there would be no way to control stopping it without
killing the instance.

This fix is essentially the same thing as before, only slightly more strict.