SimpleLibraryActivity: Refactor initialization and better debugging.
This changes the startActivityForResult to return a result so that
getCallingActivity() will work.

This fixes a bug where if a user had the popup open then another client cleared
the playlist (or the playlist cleared due to consume), THEN the button was hit
with a null current song, a crash would occur.