MPD: Fix album list with either empty list or invalid list response.
We've been getting automated crash reports where the album key is not included
in a list response line, this is probably due to an empty list or possibly an
invalid server response line. During repair, I noticed we're doing a lot of
unnecessary work with regard to album listing, so this patch fixes that at the
same time.