Tools: Refactor parseResponse() to modify in place and allow multiple arguments.
Refactor parseResponse to modify in place using a ListIterator, this will allow
the list to be modified in place and is actually a pretty large performance
gain in my non-scientific benchmarking.

This commit also renames the type parameter to keys and changes it to an arity
parameter to allow multiple key matches.