PhoneStateReceiver: Improve connection handling when not in memory.
During a refactoring earlier this year, it was not taken into account whether
MPDroid was in memory and connected. This led to not pausing due to MPDStatus
outputting a unknown status. This fixes that by moving some of the logic into
MPDControl, due mostly to the fact that we don't have a good blocking
connection lock, and MPDControl has a hacky one.

This also adds the necessary race preventers required since the connection may
take some time (hopefully not though!).