ResponseMap: Add a class to represent a response mapped skeleton.
This abstract class creates a thread-safe map of MPD key/value responses with
helper methods and fields which can be extended as necessary.

A few advantages to this approach:

* Far fewer fields per class.

* Value processing is done as necessary.

* No need to process and store every response value, to be used or not.

* This class was designed with concurrency and thread safety in mind, unlike
the older classes used to implement similar results.

* This allows easy mass value resetting, duplication, modification and
comparision.

* Value processing will be done in many smaller methods, making processing
easier to read and write and encouraging more robust processing.

* Included methods accounts for all necessary exceptions to convert to
primatives and Objects, and handles errors as required.