IdleSubsystemMonitor: Utilize the immutable status object for status comparison.
This fixes a bug that has been in the status monitor since the dmix project.
The bug affected all old states falling out of sync rarely but more often in
multiprocessing states and reconnections, which has been fixed by assigning the
old status directly prior to new status generation.

This also, obviously, moves/renames the class to the status module.