MPDCommand: Remove non-retryable commands.
There should be no case where a command should be non-retryable. The only
theory for this is probably that a command is sent, response isn't received
(because bad connection) but this is a corner case that just won't happen
and even if it WERE to happen I would expect it to be so amazingly rare
that this hack shouldn't be warranted. I can, pretty safely, say this
cannot happen with connections being what they are; there are tons of
other things that would happen first.