Layout fixes
Chiming in to do quick maintenance to at least make usuable MPDroid betas. MPDroid will not look as nice on KitKat than it does on lollipop, but that's not the goal. I won't spend hours making pre 5.0 look perfect.

 - Fix Z order on pre 5.0
 - Fix Text style (use material design standards)
 - Make list items a little smaller (72 -> 68 dp)

Some issues remain, like the default list item style that doesn't look the same in < 5.0 and 5.0.

Lots of stuff is still broken... Have fun!