CommandResponse: Update for extension and flexibility.
This commit is a radical rethinking of how we intrepret command responses. With
this commit, we take everything off the wire directly as a StringBuilder,
convert it to a string, then processing should occur directly from
the CommandResponse.

This should allow for tons of flexibility, and adds useful iterators and
methods for processing responses.

Split CommandQueues are not yet directly supported by this code.