SearchActivity: Prevent a race between the PagerView / MenuItem callback.
This commit should prevent a potential race between the PagerView field
(mPager) and the MenuItem callback. The race can present itself while the menu
is loading (menu item transition animation is active) and the user uses a
second touch to change the view, the PagerView field can be correct, while the
menu item callback can pertain to the previous view. This validates the view
for equality prior to processing the menu item click.