dmix: Refactor for Playback use.
For as long as MPDroid has been around, we've opened one thread (usually
explicit) to open another (in the connection ExecutorService) for controls.
Now that we have better future and response routines, we can remove the
explicit thread, and any required error handling will have to be done
in a thread (when it gets implemented).