MPDConnection: If status changed from connected during runnable, don't finish.
During the IOCommandProcessor exection, the status change to connected upon
the established connection. Next, the initializing connected callback is
called and will wait on the IOCommandProcessor task completion. If the
IOCommandProcessor results are retrieved (with get() and there is an
IOException, there should only be an RuntimeException generated in the case
that we are still in a "Connected" connection status state. The Callable should
have called another connection status state prior to this.