cover: Correctly encode all URLs for HTTP GET.
This patch reimplements our URI encoding for HTTP GET URLs. This comes after
continuing to get cover URL errors, even after a recent attempt to fix it with
commit d262066. After doing some research[1] I learned that it has, in fact,
been wrong for some time. This commit removes one of the queries for the send
which should make cover retrieval quicker.

This commit also changes to the preferred HTTPS scheme, when possible.

This commit also fixes the completely busted MusicBrainzCover fetching.

1. http://stackoverflow.com/a/4571518