Directory: Rewrite.
This commit rewrites directory to use the new tools provided in JMPDComm.

This commit also changes FSFragment parent directory behaviour to use the
backstack.

This commit also moves the StringComparator work to the MPDroid code,
abstracted out nicely.

This commit also adds a similar class "Listing" which is used to list all files,
including those not recognized by the server.